Move config transform into server-side load
